server: cleaned up config
